+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Add grid placement mode.
 - Add angle locking.
 - Change overlay style to match new game style.
+- Add sound effects.
 
 ### 0.9.8.6
 - Update for game version 1.2.0.